+++ src/server/js/index.js
@@ -9,4 +9,3 @@ import './event-emitter.js';
 import './functions.js';
 import './methods.js';
 import './publish.js';
-import './users-accounts.js';